+++ packages/assignment/src/App.tsx
@@ -1,151 +1,14 @@
-import React, { useState } from "react";
-import { renderLog } from "@/utils";
-import Header from "@/components/Header";
-import {
-  useItemsContext,
-  useNotificationContext,
-  useThemeContext,
-  AppProvider,
-} from "@/context";
+import React from "react";
 
-export const ItemList: React.FC = () => {
-  renderLog("ItemList rendered");
-  const { theme } = useThemeContext();
-  const { filter, filteredItems, averagePrice, handleSetFilter } =
-    useItemsContext();
+import Header from "@/app/components/Header";
+import { AppProvider } from "@/app/context";
+import { ProductsPage } from "@/pages";
 
-  return (
-    <div className="mt-8">
-      <h2 className="text-2xl font-bold mb-4">상품 목록</h2>
-      <input
-        type="text"
-        placeholder="상품 검색..."
-        value={filter}
-        onChange={(e) => handleSetFilter(e.target.value)}
-        className="w-full p-2 mb-4 border border-gray-300 rounded text-black"
-      />
-      <p className="mb-4">평균 가격: {averagePrice.toLocaleString()}원</p>
-      <ul className="space-y-2">
-        {filteredItems.slice(0, 100).map((item) => (
-          <li
-            key={item.id}
-            className={`p-2 rounded shadow ${
-              theme === "light"
-                ? "bg-white text-black"
-                : "bg-gray-700 text-white"
-            }`}
-          >
-            {item.name} - {item.category} - {item.price.toLocaleString()}원
-          </li>
-        ))}
-      </ul>
-      {filteredItems.length > 100 && (
-        <p className="mt-4">...그 외 {filteredItems.length - 100}개 상품</p>
-      )}
-    </div>
-  );
-};
-
-// ComplexForm 컴포넌트
-export const ComplexForm: React.FC = () => {
-  renderLog("ComplexForm rendered");
-  const { addNotification } = useNotificationContext();
-  const [formData, setFormData] = useState({
-    name: "",
-    email: "",
-    age: 0,
-    preferences: [] as string[],
-  });
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    addNotification("폼이 성공적으로 제출되었습니다", "success");
-  };
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({
-      ...prev,
-      [name]: name === "age" ? parseInt(value) || 0 : value,
-    }));
-  };
-
-  const handlePreferenceChange = (preference: string) => {
-    setFormData((prev) => ({
-      ...prev,
-      preferences: prev.preferences.includes(preference)
-        ? prev.preferences.filter((p) => p !== preference)
-        : [...prev.preferences, preference],
-    }));
-  };
-
-  return (
-    <div className="mt-8">
-      <h2 className="text-2xl font-bold mb-4">복잡한 폼</h2>
-      <form onSubmit={handleSubmit} className="space-y-4">
-        <input
-          type="text"
-          name="name"
-          value={formData.name}
-          onChange={handleInputChange}
-          placeholder="이름"
-          className="w-full p-2 border border-gray-300 rounded text-black"
-        />
-        <input
-          type="email"
-          name="email"
-          value={formData.email}
-          onChange={handleInputChange}
-          placeholder="이메일"
-          className="w-full p-2 border border-gray-300 rounded text-black"
-        />
-        <input
-          type="number"
-          name="age"
-          value={formData.age}
-          onChange={handleInputChange}
-          placeholder="나이"
-          className="w-full p-2 border border-gray-300 rounded text-black"
-        />
-        <div className="space-x-4">
-          {["독서", "운동", "음악", "여행"].map((pref) => (
-            <label key={pref} className="inline-flex items-center">
-              <input
-                type="checkbox"
-                checked={formData.preferences.includes(pref)}
-                onChange={() => handlePreferenceChange(pref)}
-                className="form-checkbox h-5 w-5 text-blue-600"
-              />
-              <span className="ml-2">{pref}</span>
-            </label>
-          ))}
-        </div>
-        <button
-          type="submit"
-          className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
-        >
-          제출
-        </button>
-      </form>
-    </div>
-  );
-};
-
-// 메인 App 컴포넌트
 const App: React.FC = () => {
   return (
     <AppProvider>
       <Header />
-      <div className="container mx-auto px-4 py-8">
-        <div className="flex flex-col md:flex-row">
-          <div className="w-full md:w-1/2 md:pr-4">
-            <ItemList />
-          </div>
-          <div className="w-full md:w-1/2 md:pl-4">
-            <ComplexForm />
-          </div>
-        </div>
-      </div>
+      <ProductsPage />
     </AppProvider>
   );
 };
