+++ packages/assignment/src/app/components/Header.tsx
@@ -1,6 +1,6 @@
-import { useCallback } from "../@lib";
-import { useThemeContext, useUserContext } from "../context";
-import { renderLog } from "../utils";
+import { useCallback } from "@/@lib";
+import { useThemeContext, useUserContext } from "@/app/context";
+import { renderLog } from "@/utils";
 
 // Header 컴포넌트
 const Header: React.FC = () => {
