+++ packages/assignment/src/app/context/AppProvider.tsx
@@ -1,20 +1,10 @@
-import {
-  NotificationProvider,
-  ThemeProvider,
-  UserProvider,
-  ItemsProvider,
-} from "./providers";
+import { NotificationProvider, ThemeProvider, UserProvider } from "./providers";
 
 type AppProviderProps = {
   children: React.ReactNode;
 };
 
-const providers = [
-  NotificationProvider,
-  ThemeProvider,
-  UserProvider,
-  ItemsProvider,
-];
+const providers = [NotificationProvider, ThemeProvider, UserProvider];
 
 export function AppProvider({ children }: AppProviderProps) {
   return providers.reduceRight((acc, Provider) => {
