+++ packages/assignment/src/app/context/index.ts
@@ -2,4 +2,3 @@ export { AppProvider } from "./AppProvider";
 export { useNotificationContext } from "./providers/notification/useNotificationContext";
 export { useThemeContext } from "./providers/theme/useThemeContext";
 export { useUserContext } from "./providers/user/useUserContext";
-export { useItemsContext } from "./providers/items/useItemsContext";
