+++ packages/assignment/src/app/context/providers/index.ts
@@ -1,4 +1,3 @@
 export { NotificationProvider } from "./notification/NotificationProvider";
 export { ThemeProvider } from "./theme/ThemeProvider";
 export { UserProvider } from "./user/UserProvider";
-export { ItemsProvider } from "./items/ItemsProvider";
