+++ packages/assignment/src/app/context/providers/user/UserProvider.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useCallback, useMemo } from "@/@lib";
-import { useNotificationContext } from "@/context/providers/notification/useNotificationContext";
+import { useNotificationContext } from "@/app/context/providers/notification/useNotificationContext";
 import { UserContext, UserContextType } from "./UserContext";
 
 type ThemeProviderProps = {
