--- packages/assignment/src/context/providers/items/useItemsContext.ts
@@ -1,12 +1,0 @@
-import { useContext } from "react";
-import { ItemsContext } from "./ItemsContext";
-
-export const useItemsContext = () => {
-  const context = useContext(ItemsContext);
-  if (!context) {
-    throw new Error(
-      "useItemsContext는 ItemsProvider 내부에서 사용해야 합니다."
-    );
-  }
-  return context;
-};
