+++ packages/assignment/src/pages/index.ts
@@ -1,0 +1,1 @@
+export { default as ProductsPage } from "./products/ProductsPage";
