+++ packages/assignment/src/pages/products/components/index.ts
@@ -1,0 +1,2 @@
+export { default as ProductPanel } from "./ProductPanel";
+export { default as ProductAddForm } from "./ProductAddForm";
